requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==44.0.3
daphne==4.1.2
Django==5.2.1
django-cors-headers==4.7.0
django-decouple==2.1
django-filter==25.1
djangorestframework==3.16.0
drf-spectacular==0.28.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
Markdown==3.8
msgpack==1.1.0
pillow==11.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.0.0
PyYAML==6.0.2
redis==6.0.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
service-identity==24.2.0
sqlparse==0.5.3
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
whitenoise==6.9.0
zope.interface==7.2
